Move Proto field related classes to `base` package
